pyproject.toml
[project]
name = "polly"
version = "0.1.0"
description = "Discord poll bot with web interface"
authors = [{name = "Polly Team"}]
requires-python = ">=3.11"
dependencies = [
    "discord.py>=2.4.0",
    "fastapi>=0.115.0",
    "uvicorn>=0.32.0",
    "sqlalchemy>=2.0.36",
    "aiosqlite>=0.20.0",
    "apscheduler>=3.11.0",
    "jinja2>=3.1.4",
    "python-multipart>=0.0.12",
    "httpx>=0.28.0",
    "python-jose[cryptography]>=3.3.0",
    "python-dotenv>=1.0.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = []
